requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
langchain
langchain-community
chromadb
sentence-transformers
python-dotenv
pydantic